Still outrageously hacky and riddled with races, but can sort-of handle completion requests
