--- ycmd/completers/java/server_stderr
@@ -1,32 +1,0 @@
-_RegisterApplication(), FAILED TO establish the default connection to the WindowServer, _CGSDefaultConnection() is NULL.
-SLF4J: Failed to load class "org.slf4j.impl.StaticLoggerBinder".
-SLF4J: Defaulting to no-operation (NOP) logger implementation
-SLF4J: See http://www.slf4j.org/codes.html#StaticLoggerBinder for further details.
-java.lang.RuntimeException: java.net.SocketException: Broken pipe
-	at org.eclipse.lsp4j.jsonrpc.json.StreamMessageConsumer.consume(StreamMessageConsumer.java:68)
-	at org.eclipse.lsp4j.jsonrpc.RemoteEndpoint.notify(RemoteEndpoint.java:84)
-	at org.eclipse.lsp4j.jsonrpc.services.EndpointProxy.invoke(EndpointProxy.java:61)
-	at com.sun.proxy.$Proxy5.logMessage(Unknown Source)
-	at org.jboss.tools.vscode.java.internal.JavaClientConnection.logMessage(JavaClientConnection.java:49)
-	at org.jboss.tools.vscode.java.internal.handlers.LogHandler.processLogMessage(LogHandler.java:97)
-	at org.jboss.tools.vscode.java.internal.handlers.LogHandler.access$0(LogHandler.java:85)
-	at org.jboss.tools.vscode.java.internal.handlers.LogHandler$1.logging(LogHandler.java:54)
-	at org.eclipse.core.internal.runtime.RuntimeLog.logToListeners(RuntimeLog.java:161)
-	at org.eclipse.core.internal.runtime.PlatformLogWriter.logged(PlatformLogWriter.java:103)
-	at org.eclipse.osgi.internal.log.ExtendedLogReaderServiceFactory.safeLogged(ExtendedLogReaderServiceFactory.java:88)
-	at org.eclipse.osgi.internal.log.ExtendedLogReaderServiceFactory.logPrivileged(ExtendedLogReaderServiceFactory.java:217)
-	at org.eclipse.osgi.internal.log.ExtendedLogReaderServiceFactory.log(ExtendedLogReaderServiceFactory.java:189)
-	at org.eclipse.osgi.internal.log.ExtendedLogServiceFactory.log(ExtendedLogServiceFactory.java:65)
-	at org.eclipse.osgi.internal.log.ExtendedLogServiceImpl.log(ExtendedLogServiceImpl.java:87)
-	at org.eclipse.osgi.internal.log.LoggerImpl.log(LoggerImpl.java:54)
-	at org.eclipse.core.internal.runtime.PlatformLogWriter.logging(PlatformLogWriter.java:44)
-	at org.eclipse.core.internal.runtime.RuntimeLog.log(RuntimeLog.java:97)
-	at org.eclipse.core.internal.jobs.JobManager.endJob(JobManager.java:718)
-	at org.eclipse.core.internal.jobs.WorkerPool.endJob(WorkerPool.java:105)
-	at org.eclipse.core.internal.jobs.Worker.run(Worker.java:72)
-Caused by: java.net.SocketException: Broken pipe
-	at java.net.SocketOutputStream.socketWrite0(Native Method)
-	at java.net.SocketOutputStream.socketWrite(SocketOutputStream.java:109)
-	at java.net.SocketOutputStream.write(SocketOutputStream.java:141)
-	at org.eclipse.lsp4j.jsonrpc.json.StreamMessageConsumer.consume(StreamMessageConsumer.java:63)
-	... 20 more
